common-requirements.txt
#####
# Requirements needed on all machines running Heroku
#
# Due to issues with some packages compiling on dev machines we break out
# requirements that are needed on Heroku (requirements.txt) and on dev machines
# (dev-requirements.txt)
Django==1.8.5
psycopg2==2.6

# Environment Manager
django-environ==0.3.0


# Production Operations Packages
newrelic==2.52.0.40


# Core Requirements
beautifulsoup4==4.3.2
django-autocomplete-light==2.1.1
django-celery==3.1.16
django-extra-views==0.6.5
django-ipware==1.0.0
django-pure-pagination==0.2.1
django-taggit==0.14.0
django-widget-tweaks==1.3
django_extensions==1.5.5
flanker==0.4.29
geopy==1.10.0
jsonfield==1.0.3
lxml==3.4.4
Pillow==2.8.1
pytz==2013b
requests==2.7.0
simplejson==3.7.1
tablib==0.10.0
ua-parser==0.3.6
unidecode==0.04.17


# AWS Specific Packages
boto==2.38.0


# Simple Email Service Packages
django-bouncy==0.0.1
Twisted==15.4.0
service_identity==14.0.0
seacucumber==1.5.2


# S3/Storage Packages
django-storages==1.1.8
django-s3-collectstatic==1.0
django-s3-folder-storage==0.2


# Testing Related Packages
mock==1.0.1
model-mommy==1.2.5
nose==1.3.6
django-nose==1.4.0
nose_timer==0.4.4
dj-inmemorystorage==1.3.0
coverage==3.7.1
nosexcover==1.0.10

# Useful packages for Debugging
django-clear-cache==0.3
django-debug-toolbar==1.3.0
ipython==3.1.0
readline==6.2.4.1

# Authentication Libraries
-e git+https://github.com/ofa/python-social-auth.git@ngp-actionid#egg=python-social-auth
